web/requirements.txt
Flask==3.0.0
grpcio==1.60.0
protobuf==4.25.0
pytest==7.4.0
requests==2.31.0